requirements.txt
discord.py
requests
flask
python-dotenv
firebase-admin